@@ -13,47 +13,47 @@ extension GetBranchCollection on Isar {
   IsarCollection<Branch> get branchs => this.collection();
 }
 
-final BranchSchema = CollectionSchema(
+const BranchSchema = CollectionSchema(
   name: r'Branch',
-  id: BigInt.parse("7089804700259644623").toInt(),
+  id: 7089804700259644623,
   properties: {
     r'active': PropertySchema(
-      id: BigInt.parse("0").toInt(),
+      id: 0,
       name: r'active',
       type: IsarType.bool,
     ),
     r'businessId': PropertySchema(
-      id: BigInt.parse("1").toInt(),
+      id: 1,
       name: r'businessId',
       type: IsarType.long,
     ),
     r'description': PropertySchema(
-      id: BigInt.parse("2").toInt(),
+      id: 2,
       name: r'description',
       type: IsarType.string,
     ),
     r'isDefault': PropertySchema(
-      id: BigInt.parse("3").toInt(),
+      id: 3,
       name: r'isDefault',
       type: IsarType.bool,
     ),
     r'latitude': PropertySchema(
-      id: BigInt.parse("4").toInt(),
+      id: 4,
       name: r'latitude',
       type: IsarType.string,
     ),
     r'longitude': PropertySchema(
-      id: BigInt.parse("5").toInt(),
+      id: 5,
       name: r'longitude',
       type: IsarType.string,
     ),
     r'name': PropertySchema(
-      id: BigInt.parse("6").toInt(),
+      id: 6,
       name: r'name',
       type: IsarType.string,
     ),
     r'table': PropertySchema(
-      id: BigInt.parse("7").toInt(),
+      id: 7,
       name: r'table',
       type: IsarType.string,
     )
@@ -69,7 +69,7 @@ final BranchSchema = CollectionSchema(
   getId: _branchGetId,
   getLinks: _branchGetLinks,
   attach: _branchAttach,
-  version: '3.1.0',
+  version: '3.1.0+1',
 );
 
 int _branchEstimateSize(
