@@ -13,202 +13,202 @@ extension GetBusinessCollection on Isar {
   IsarCollection<Business> get business => this.collection();
 }
 
-final BusinessSchema = CollectionSchema(
+const BusinessSchema = CollectionSchema(
   name: r'Business',
-  id: BigInt.parse("-8587991543652827682").toInt(),
+  id: -8587991543652827682,
   properties: {
     r'active': PropertySchema(
-      id: BigInt.parse("0").toInt(),
+      id: 0,
       name: r'active',
       type: IsarType.bool,
     ),
     r'adrs': PropertySchema(
-      id: BigInt.parse("1").toInt(),
+      id: 1,
       name: r'adrs',
       type: IsarType.string,
     ),
     r'backUpEnabled': PropertySchema(
-      id: BigInt.parse("2").toInt(),
+      id: 2,
       name: r'backUpEnabled',
       type: IsarType.bool,
     ),
     r'backupFileId': PropertySchema(
-      id: BigInt.parse("3").toInt(),
+      id: 3,
       name: r'backupFileId',
       type: IsarType.string,
     ),
     r'bhfId': PropertySchema(
-      id: BigInt.parse("4").toInt(),
+      id: 4,
       name: r'bhfId',
       type: IsarType.string,
     ),
     r'businessTypeId': PropertySchema(
-      id: BigInt.parse("5").toInt(),
+      id: 5,
       name: r'businessTypeId',
       type: IsarType.long,
     ),
     r'businessUrl': PropertySchema(
-      id: BigInt.parse("6").toInt(),
+      id: 6,
       name: r'businessUrl',
       type: IsarType.string,
     ),
     r'categoryId': PropertySchema(
-      id: BigInt.parse("7").toInt(),
+      id: 7,
       name: r'categoryId',
       type: IsarType.string,
     ),
     r'channels': PropertySchema(
-      id: BigInt.parse("8").toInt(),
+      id: 8,
       name: r'channels',
       type: IsarType.stringList,
     ),
     r'chatUid': PropertySchema(
-      id: BigInt.parse("9").toInt(),
+      id: 9,
       name: r'chatUid',
       type: IsarType.string,
     ),
     r'country': PropertySchema(
-      id: BigInt.parse("10").toInt(),
+      id: 10,
       name: r'country',
       type: IsarType.string,
     ),
     r'createdAt': PropertySchema(
-      id: BigInt.parse("11").toInt(),
+      id: 11,
       name: r'createdAt',
       type: IsarType.string,
     ),
     r'currency': PropertySchema(
-      id: BigInt.parse("12").toInt(),
+      id: 12,
       name: r'currency',
       type: IsarType.string,
     ),
     r'deviceToken': PropertySchema(
-      id: BigInt.parse("13").toInt(),
+      id: 13,
       name: r'deviceToken',
       type: IsarType.string,
     ),
     r'dvcSrlNo': PropertySchema(
-      id: BigInt.parse("14").toInt(),
+      id: 14,
       name: r'dvcSrlNo',
       type: IsarType.string,
     ),
     r'email': PropertySchema(
-      id: BigInt.parse("15").toInt(),
+      id: 15,
       name: r'email',
       type: IsarType.string,
     ),
     r'firstName': PropertySchema(
-      id: BigInt.parse("16").toInt(),
+      id: 16,
       name: r'firstName',
       type: IsarType.string,
     ),
     r'fullName': PropertySchema(
-      id: BigInt.parse("17").toInt(),
+      id: 17,
       name: r'fullName',
       type: IsarType.string,
     ),
     r'hexColor': PropertySchema(
-      id: BigInt.parse("18").toInt(),
+      id: 18,
       name: r'hexColor',
       type: IsarType.string,
     ),
     r'imageUrl': PropertySchema(
-      id: BigInt.parse("19").toInt(),
+      id: 19,
       name: r'imageUrl',
       type: IsarType.string,
     ),
     r'isDefault': PropertySchema(
-      id: BigInt.parse("20").toInt(),
+      id: 20,
       name: r'isDefault',
       type: IsarType.bool,
     ),
     r'isLastSubscriptionPaymentSucceeded': PropertySchema(
-      id: BigInt.parse("21").toInt(),
+      id: 21,
       name: r'isLastSubscriptionPaymentSucceeded',
       type: IsarType.bool,
     ),
     r'lastDbBackup': PropertySchema(
-      id: BigInt.parse("22").toInt(),
+      id: 22,
       name: r'lastDbBackup',
       type: IsarType.string,
     ),
     r'lastName': PropertySchema(
-      id: BigInt.parse("23").toInt(),
+      id: 23,
       name: r'lastName',
       type: IsarType.string,
     ),
     r'lastSeen': PropertySchema(
-      id: BigInt.parse("24").toInt(),
+      id: 24,
       name: r'lastSeen',
       type: IsarType.long,
     ),
     r'latitude': PropertySchema(
-      id: BigInt.parse("25").toInt(),
+      id: 25,
       name: r'latitude',
       type: IsarType.string,
     ),
     r'longitude': PropertySchema(
-      id: BigInt.parse("26").toInt(),
+      id: 26,
       name: r'longitude',
       type: IsarType.string,
     ),
     r'metadata': PropertySchema(
-      id: BigInt.parse("27").toInt(),
+      id: 27,
       name: r'metadata',
       type: IsarType.string,
     ),
     r'name': PropertySchema(
-      id: BigInt.parse("28").toInt(),
+      id: 28,
       name: r'name',
       type: IsarType.string,
     ),
     r'nextBillingDate': PropertySchema(
-      id: BigInt.parse("29").toInt(),
+      id: 29,
       name: r'nextBillingDate',
       type: IsarType.string,
     ),
     r'previousBillingDate': PropertySchema(
-      id: BigInt.parse("30").toInt(),
+      id: 30,
       name: r'previousBillingDate',
       type: IsarType.string,
     ),
     r'role': PropertySchema(
-      id: BigInt.parse("31").toInt(),
+      id: 31,
       name: r'role',
       type: IsarType.string,
     ),
     r'subscriptionPlan': PropertySchema(
-      id: BigInt.parse("32").toInt(),
+      id: 32,
       name: r'subscriptionPlan',
       type: IsarType.string,
     ),
     r'taxEnabled': PropertySchema(
-      id: BigInt.parse("33").toInt(),
+      id: 33,
       name: r'taxEnabled',
       type: IsarType.bool,
     ),
     r'taxServerUrl': PropertySchema(
-      id: BigInt.parse("34").toInt(),
+      id: 34,
       name: r'taxServerUrl',
       type: IsarType.string,
     ),
     r'timeZone': PropertySchema(
-      id: BigInt.parse("35").toInt(),
+      id: 35,
       name: r'timeZone',
       type: IsarType.string,
     ),
     r'tinNumber': PropertySchema(
-      id: BigInt.parse("36").toInt(),
+      id: 36,
       name: r'tinNumber',
       type: IsarType.long,
     ),
     r'type': PropertySchema(
-      id: BigInt.parse("37").toInt(),
+      id: 37,
       name: r'type',
       type: IsarType.string,
     ),
     r'userId': PropertySchema(
-      id: BigInt.parse("38").toInt(),
+      id: 38,
       name: r'userId',
       type: IsarType.string,
     )
@@ -220,7 +220,7 @@ final BusinessSchema = CollectionSchema(
   idName: r'id',
   indexes: {
     r'userId': IndexSchema(
-      id: BigInt.parse("-2005826577402374815").toInt(),
+      id: -2005826577402374815,
       name: r'userId',
       unique: false,
       replace: false,
@@ -238,7 +238,7 @@ final BusinessSchema = CollectionSchema(
   getId: _businessGetId,
   getLinks: _businessGetLinks,
   attach: _businessAttach,
-  version: '3.1.0',
+  version: '3.1.0+1',
 );
 
 int _businessEstimateSize(
