@@ -13,12 +13,12 @@ extension GetPermissionCollection on Isar {
   IsarCollection<Permission> get permissions => this.collection();
 }
 
-final PermissionSchema = CollectionSchema(
+const PermissionSchema = CollectionSchema(
   name: r'Permission',
-  id: BigInt.parse("5583744244740860356").toInt(),
+  id: 5583744244740860356,
   properties: {
     r'name': PropertySchema(
-      id: BigInt.parse("0").toInt(),
+      id: 0,
       name: r'name',
       type: IsarType.string,
     )
@@ -34,7 +34,7 @@ final PermissionSchema = CollectionSchema(
   getId: _permissionGetId,
   getLinks: _permissionGetLinks,
   attach: _permissionAttach,
-  version: '3.1.0',
+  version: '3.1.0+1',
 );
 
 int _permissionEstimateSize(
