@@ -13,37 +13,37 @@ extension GetVoucherCollection on Isar {
   IsarCollection<Voucher> get vouchers => this.collection();
 }
 
-final VoucherSchema = CollectionSchema(
+const VoucherSchema = CollectionSchema(
   name: r'Voucher',
-  id: BigInt.parse("6548118931563840322").toInt(),
+  id: 6548118931563840322,
   properties: {
     r'createdAt': PropertySchema(
-      id: BigInt.parse("0").toInt(),
+      id: 0,
       name: r'createdAt',
       type: IsarType.long,
     ),
     r'descriptor': PropertySchema(
-      id: BigInt.parse("1").toInt(),
+      id: 1,
       name: r'descriptor',
       type: IsarType.string,
     ),
     r'interval': PropertySchema(
-      id: BigInt.parse("2").toInt(),
+      id: 2,
       name: r'interval',
       type: IsarType.long,
     ),
     r'used': PropertySchema(
-      id: BigInt.parse("3").toInt(),
+      id: 3,
       name: r'used',
       type: IsarType.bool,
     ),
     r'usedAt': PropertySchema(
-      id: BigInt.parse("4").toInt(),
+      id: 4,
       name: r'usedAt',
       type: IsarType.long,
     ),
     r'value': PropertySchema(
-      id: BigInt.parse("5").toInt(),
+      id: 5,
       name: r'value',
       type: IsarType.long,
     )
@@ -59,7 +59,7 @@ final VoucherSchema = CollectionSchema(
   getId: _voucherGetId,
   getLinks: _voucherGetLinks,
   attach: _voucherAttach,
-  version: '3.1.0',
+  version: '3.1.0+1',
 );
 
 int _voucherEstimateSize(
