@@ -1,21 +1,21 @@
-import 'package:injectable/injectable.dart';
+// import 'package:injectable/injectable.dart';
 
-import 'api.dart';
-import 'service.dart';
-import 'storage.dart';
+// import 'api.dart';
+// import 'service.dart';
+// import 'storage.dart';
 
-@module
-abstract class Module {
-  @LazySingleton()
-  Storage box() {
-    return StorageImpl();
-  }
+// @module
+// abstract class Module {
+//   @LazySingleton()
+//   Storage box() {
+//     return StorageImpl();
+//   }
 
-  @preResolve
-  @LazySingleton()
-  Future<AB> api() async {
-    return Api();
-  }
+//   @preResolve
+//   @LazySingleton()
+//   Future<AB> api() async {
+//     return Api();
+//   }
 
-  Service get service;
-}
+//   Service get service;
+// }
